Adds ability to read LesHouches GiBUU output.
  - Makes the FinalEvents.dat parser more robust as well.
  - Begins to add the ability to confidently read in initial
    state information if present.

+++ GiBUUToStdHep/GiBUUToStdHep_Utils.cxx
@@ -50,7 +50,7 @@ std::string int2str(int i){
   return ss.str();
 }
 
-int GiBUUToPDG(int GiBUUCode, double GiBUUCharge){
+int GiBUUToPDG(int GiBUUCode, int GiBUUCharge){
   //https://gibuu.hepforge.org/trac/wiki/ParticleIDs
   switch(GiBUUCode){
     case 1:{
